audio: Avoid mute during AFE PROXY Usecase
Silence data is getting written from in_read when voice
call is muted with USB Headset connected. This resulting
in muting Rx path too along with Tx path in this
scenario. USB Voice call scenario, PROXY_IN device is
responsible for reading RX data. So not supposed to be
muted.

Fix is to avoid muting if the current use case is AFE_PROXY

CRs-Fixed: 763152

Change-Id: I1fbd5545f343cbaea865e941214be05791dec14e